reorganized files and updated schema
