+++ authenticate-me/backend/db/models/user.js
@@ -60,6 +60,14 @@ module.exports = (sequelize, DataTypes) => {
       }
 
     },
+    firstName: {
+      allowNull: false,
+      type: DataTypes.STRING
+    },
+    lastName: {
+      allowNull: false,
+      type: DataTypes.STRING
+    },
     email: {
       allowNull: false,
       type: DataTypes.STRING,
