--- authenticate-me/backend/routes/api/index.js
@@ -1,15 +1,0 @@
-const  router = require('express').Router();
-const sessionRouter = require('./session');
-const usersRouter = require('./users');
-
-router.use('/session', sessionRouter);
-router.use('/users', usersRouter);
-
-router.post('/test', (req, res) => {
-    res.json({ requestBody: req.body });
-});
-
-
-
-
-module.exports = router;
