+++ authenticate-me/backend/routes/login.js
@@ -1,8 +1,8 @@
 const express = require('express');
-const  { setTokenCookie, restoreUser } = require('../../utils/auth');
-const { User } = require('../../db/models');
+const  { setTokenCookie, restoreUser } = require('../utils/auth');
+const { User } = require('../db/models');
 const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+const { handleValidationErrors } = require('../utils/validation');
 
 const router = express.Router();
