+++ authenticate-me/backend/routes/users.js
@@ -1,8 +1,8 @@
 const express =  require('express');
-const { setTokenCookie, requireAuth } = require('../../utils/auth');
-const { User } = require('../../db/models');
+const { setTokenCookie, requireAuth } = require('../utils/auth');
+const { User } = require('../db/models');
 const { check } = require('express-validator');
-const { handleValidationErrors } = require('../../utils/validation');
+const { handleValidationErrors } = require('../utils/validation');
 
 const router = express.Router();
 
@@ -28,7 +28,7 @@ const validateSignup = [
   ];
 
 //Sign up
-router.post('/', validateSignup, async (req, res) => {
+router.post('/signup', validateSignup, async (req, res) => {
     const { email, password, username } = req.body;
 
     const user = await User.signup({ email, username, password });
@@ -38,6 +38,11 @@ router.post('/', validateSignup, async (req, res) => {
     return res.json({
         user
     });
+});
+
+//Get Current User
+router.get('/current', async(req, res) => {
+
 });
